feat : 불필요한 코드 삭제 및 폴더구조 변경
