--- nonsoolmateClient/src/socialLogin/api/getData.ts
@@ -1,7 +1,0 @@
-import { useQueryClient } from "react-query";
-
-export function getData() {
-  const queryClient = useQueryClient();
-
-  console.log(queryClient.getQueriesData);
-}
