--- nonsoolmateClient/src/socialLogin/hooks/type.ts
@@ -1,13 +1,0 @@
-export interface loginResProps {
-  data: {
-    code: number;
-    message: string;
-    data: {
-      memberId: number;
-      authType: string;
-      memberName: string;
-      accessToken: string;
-      refreshToken: string;
-    };
-  };
-}
